Moved Segment code to its own file
